refactor(report): move report risk helpers into utils folder
